% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDSinRdata-package.R
\docType{package}
\name{HDSinRdata-package}
\alias{HDSinRdata}
\alias{HDSinRdata-package}
\title{HDSinRdata: Data Package for the 'Health Data Science in R' Online Textbook}
\description{
Contains seven datasets used in the chapters and exercises of the 'Health Data Science in R' Online Textbook.
}
\author{
\strong{Maintainer}: Joanna Walsh \email{joanna_walsh@brown.edu}

Authors:
\itemize{
  \item Alice Paul \email{alice_paul@brown.edu}
}

Other contributors:
\itemize{
  \item Jialin Liu [contributor]
  \item Xinbei Yu [contributor]
}

}
\keyword{internal}
