% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RforHDSdata-package.R
\docType{package}
\name{RforHDSdata-package}
\alias{RforHDSdata}
\alias{RforHDSdata-package}
\title{RforHDSdata: Data Package for the 'R for Health Data Science' Online Textbook}
\description{
This package contains seven datasets used in the chapters and exercises of the 'R for Health Data Science' Online Textbook.
}
\author{
\strong{Maintainer}: Joanna Walsh \email{joanna_walsh@brown.edu}

Authors:
\itemize{
  \item Alice Paul \email{alice_paul@brown.edu}
}

Other contributors:
\itemize{
  \item Jialin Liu [contributor]
  \item Xinbei Yu [contributor]
}

}
\keyword{internal}
