% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDSinRdata-package.R
\docType{package}
\name{HDSinRdata-package}
\alias{HDSinRdata}
\alias{HDSinRdata-package}
\title{HDSinRdata: Data for the 'Mastering Health Data Science Using R' Online Textbook}
\description{
Contains ten datasets used in the chapters and exercises of Paul, Alice (2023) "Health Data Science in R" \url{https://alicepaul.github.io/health-data-science-using-r/}.
}
\author{
\strong{Maintainer}: Alice Paul \email{alice_paul@brown.edu}

Authors:
\itemize{
  \item Hannah Eglinton \email{hannah_eglinton@brown.edu}
  \item Joanna Walsh \email{joanna_walsh@brown.edu}
}

Other contributors:
\itemize{
  \item Jialin Liu [contributor]
  \item Xinbei Yu [contributor]
}

}
\keyword{internal}
