% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockdowndates.R
\docType{data}
\name{lockdowndates}
\alias{lockdowndates}
\title{Lockdown dates from Ballotpedia}
\format{
A data frame with 50 rows and 3 variables:
\describe{
  \item{State}{State}
  \item{Lockdown_Start}{Start date of the statewide order in YYYY-MM-DD
  format}
  \item{Lockdown_End}{End date of the statewide order in YYYY-MM-DD
  format}
}
}
\source{
Raifman, J., Nocka, K., Jones, D., Bor, J., Lipson, S., 
Jay, J., Cole, M., Krawczyk, N., Benfer, E. A.,  Chan, P.,  Galea, S. (2022). 
COVID-19 US State Policy Database. Ann Arbor, MI: 
Inter-university Consortium for Political and Social Research [distributor], 
2022-03-30. https://doi.org/10.3886/E119446V143
 
\url{https://www.openicpsr.org/openicpsr/project/119446/version/V143/view}
}
\usage{
lockdowndates
}
\description{
Start and end dates of statewide stay at home orders in response to the
Covid-19 pandemic.
}
\keyword{datasets}
